Ensure all tools are building with -Wall and -Werror
Corrected resulting compiler errors from correctly adding -Wall &
-Werror within configure.ac.

+++ storageaddsel.cpp
@@ -76,10 +76,10 @@ int find_sensor_type_string(uint8_t sensor_number, char **s) {
 }
 
 
-size_t getfilestream(const char *fn, uint8_t **buffer) {
+long getfilestream(const char *fn, uint8_t **buffer) {
 
 	FILE *fp;
-	size_t size = 0;
+	long size = 0;
 	int r;
 
 	if ((fp = fopen(fn, "rb")) != NULL) {
@@ -234,7 +234,7 @@ void send_esel(uint16_t recordid) {
 	const char *sev;
 	uint8_t *buffer = NULL;
 	const char *path = "/tmp/esel";
-	size_t sz;
+	long sz;
 	int r;
 
 	sz = getfilestream(path, &buffer);
